backend/requirements.txt
# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Web scraping and content processing
beautifulsoup4==4.12.3
requests==2.31.0
playwright==1.41.2

# AI and configuration
google-generativeai==0.3.2
python-dotenv==1.0.1 